extract RaftLog.* to a standalone module
